+++ src/app/components/main/containers/PortfolioContainer.tsx
@@ -1,6 +1,6 @@
 import { LOGGEDIN_SECTION_HEADINGS } from 'constants/main';
-import PortfolioItem from '../PortfolioItem';
-import PortfolioItemNav from '../PortfolioItemNav';
+import PortfolioItem from '../portfolio/PortfolioItem';
+import PortfolioItemNav from '../portfolio/PortfolioItemNav';
 import Pages from 'components/common/Pages';
 
 const PortfolioContainer = () => {
