+++ src/app/components/main/containers/RecruitingContainer.tsx
@@ -1,0 +1,23 @@
+import RecruitingItem from '../recruiting/RecruitingItem';
+
+interface RecruitingContainerProps {
+  user: UserData;
+  data: RecruitListItem[];
+}
+
+const RecruitingContainer = ({ user, data }: RecruitingContainerProps) => {
+  return (
+    <div className="mt-[75px]">
+      {data.map((item) => {
+        return (
+          <div key={item.articleId}>
+            {/* <Profile /> */}
+            <RecruitingItem />
+          </div>
+        );
+      })}
+    </div>
+  );
+};
+
+export default RecruitingContainer;
