+++ src/app/components/main/recruiting/RecruitingItem.tsx
@@ -1,0 +1,9 @@
+const RecruitingItem = () => {
+  return (
+    <div className="w-[1280px] h-[809px] bg-white text-black rounded-2xl">
+      RecruitingItem
+    </div>
+  );
+};
+
+export default RecruitingItem;
